+++ internal/server/contentmigrate_test.go
@@ -96,6 +96,7 @@ var notIDs = map[string]string{
 	"server.containerFile.HiveItems{}.Honey":            whyCount,
 	"server.containerFile.HiveItems{}.Occ[].Flower[][]": whyPos,
 	"server.containerFile.HiveItems{}.Occ[].SecsLeft":   whyCount,
+	"server.containerFile.Items[].Age":                  whyCount,
 	"server.containerFile.Items[].Book":                 whyRef,
 	"server.containerFile.Items[].Box":                  whyRef,
 	"server.containerFile.Items[].Bundle":               whyRef,